# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Play with Friends Stream API
  version: 1.0.0
  description: Backend API for the Snake Game multiplayer application.

servers:
  - url: http://localhost:8000
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        highScore:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - highScore
        - createdAt

    AuthCredentials:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        username:
          type: string
      required:
        - email
        - password

    LeaderboardEntry:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        score:
          type: integer
        mode:
          $ref: '#/components/schemas/GameMode'
        date:
          type: string
          format: date-time
        rank:
          type: integer
      required:
        - id
        - username
        - score
        - mode
        - date
        - rank

    GameMode:
      type: string
      enum:
        - pass-through
        - walls

    Direction:
      type: string
      enum:
        - UP
        - DOWN
        - LEFT
        - RIGHT

    GameStatus:
      type: string
      enum:
        - idle
        - playing
        - paused
        - game-over

    Position:
      type: object
      properties:
        x:
          type: integer
        y:
          type: integer
      required:
        - x
        - y

    ActivePlayer:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        score:
          type: integer
        mode:
          $ref: '#/components/schemas/GameMode'
        snake:
          type: array
          items:
            $ref: '#/components/schemas/Position'
        food:
          $ref: '#/components/schemas/Position'
        direction:
          $ref: '#/components/schemas/Direction'
        status:
          $ref: '#/components/schemas/GameStatus'
      required:
        - id
        - username
        - score
        - mode
        - snake
        - food
        - direction
        - status

paths:
  /auth/signup:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCredentials'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid input or email already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string

  /auth/login:
    post:
      summary: Login user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    description: JWT token for authentication
        '401':
          description: Invalid credentials

  /auth/logout:
    post:
      summary: Logout user
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true

  /auth/me:
    get:
      summary: Get current logged-in user
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated

  /leaderboard:
    get:
      summary: Get leaderboard entries
      tags:
        - Leaderboard
      parameters:
        - in: query
          name: mode
          schema:
            $ref: '#/components/schemas/GameMode'
          required: false
          description: Filter by game mode
      responses:
        '200':
          description: Leaderboard entries retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntry'

    post:
      summary: Submit a new score
      tags:
        - Leaderboard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                score:
                  type: integer
                mode:
                  $ref: '#/components/schemas/GameMode'
              required:
                - score
                - mode
      responses:
        '201':
          description: Score submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/LeaderboardEntry'
        '401':
          description: Unauthorized

  /active-players:
    get:
      summary: Get all active players
      tags:
        - Active Players
      responses:
        '200':
          description: List of active players
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivePlayer'

  /active-players/{id}:
    get:
      summary: Get a specific active player by ID
      tags:
        - Active Players
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the active player
      responses:
        '200':
          description: Active player details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ActivePlayer'
        '404':
          description: Player not found